requirements.txt
flask==2.3.2
pymongo==4.6.1
python-dotenv==1.0.1
bcrypt==4.1.2
pydantic==2.7.1
gunicorn==22.0.0
pytest==8.2.0